added quality options

+++ Controllers/JSONClasses/SetupQualityJsonClass.cs
@@ -1,6 +1,6 @@
 namespace voddy.Controllers.Structures {
     public class SetupQualityJsonClass {
         public int Resolution { get; set; }
-        public int Fps { get; set; }
+        public double Fps { get; set; }
     }
 }
